requirement.txt
flask
requests
cv2
matplotlib
sklearn
pickle
numpy
pandas
sklearn
keras
Pillow
io